Finish greedy naive approach
This method of finding the highest gold doesn't take into account zeroes
or any of the other starting positions. It does beat the 10th person on
the leaderboard but that's not good enough.

Next I would like to determine some of the best starting positions,
maybe start somewhere near the middle, and take into account zeroes.