Implement findZeroes to avoid zeroes
To get a better score we need to avoid times where there are two zeroes
approaching ahead. We do this by looking ahead and planning accordingly.

To make this happen we find occurences of double zeroes and make a
different legal move ahead of time. If we don't find double zeroes we
carry on as regular.