Implement dynamic programming solution
This commit adds functionality to determine a better (though sadly not
optimal) path through the mine.

Beginning with the last column, we iterate backwards through the mine,
storing the "potential" value of each node (i.e. the value that can be
achieved from a path beginning at said node, as well as the next
direction to take for that path.

Rather than the brute-force approach of trying each path (which would
result in exponential run-time), this approach reduces that to
polynomial time, specifically O(m * n), where `m` and `n` are the
dimensions of the mine. Though there are various internal operations
(e.g. sorting the available moves), the iteration represents the bulk of
the run time.

In trying to improve my solution, I recognize that there are several
limitations, chief among them the issue that the algorithm can choose
the best path in one column, only to have that path be ineligble for a
node in the next column. In this instance, the node will swing to the
alternate direction, even if it is still possible to use that "richest"
node by changing the path direction to avoid duplicate movements.