Refactor solution to improve mining efficacy
This commit refactors the solution function. Instead of simply storing
the best possible direction at each node, and then using this at the end
to determine the best path, we now store information on the _two_ best
paths possible from each node.

The general outline is the same, i.e. we iterate backwards through each
column of the mine array, however, we now store the following
information in our cache:
  * value
  * the "best path" information (next direction, potential value, and
    the path itself)
  * the "alternate path" information

The previous approach resulted in certain paths being "lost", since it
was entirely possible that, for a given node, the best path through the
richest node isn't possible (since it results in duplicate movements),
but the second best path through the richest node is still the ideal
path. This approach ensures that such scenarios are accounted for.

The paths are stored here as an array of numbers corresponding to the
direction to take; when the best possible starting node is identified,
we can simply use this corresponding array and dequeue each time
`move()` is called to get the next Y position.

This did result in a significantly larger space complexity, however, by
setting cache entries to `null` once they are no longer needed, the run
time has been reduced to a reasonable level.